android: start proximity sensor anyway
Proximity sensor should be enabled anyway, but only turn screen off when audio is
routing to earpiece while proximity is in `is near` state.

This is for app which can implement `choose audio route` functionality. If a
user start a video call but then wants to use earpiece to continue the
rest conversation, app should be able to route audio to EarPiece and
`turn screen off` while user put their ear close to screen.